k12_verify/requirements.txt
# K12 Verify - SheerID Teacher Verification Tool

## Requirements

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# HTTP Client
httpx>=0.25.0
cloudscraper>=1.2.71  # Cloudflare bypass (from military module)

# Document Generation
playwright>=1.40.0
pillow>=10.0.0

# Config
python-dotenv>=1.0.0

# WebSocket
websockets>=12.0
